+++ kernel/tests/test.rs
@@ -15,32 +15,10 @@ unsafe fn kernel_init() -> ! {
 
 
 
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-// in src/main.rs
-
-impl<T> Testable for T
-where
-    T: Fn(),
-{
-    fn run(&self) {
-        println!("{}...\t", core::any::type_name::<T>());
-        self();
-        println!("[ok]");
+#[cfg(test)]
+pub fn test_runner(tests: &[&dyn Fn()]) {
+    println!("Running {} tests", tests.len());
+    for test in tests {
+        test();
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #[test_case]
-    const TEST1: libkernel::UnitTest = libkernel::UnitTest {
-            name: "test_runner_executes_in_kernel_mode",
-            test_func: || {
-                assert!(1 + 1 == 2);
-            },
-        };
-}
-
-// SPDX-License-Identifier: MIT OR Apache-2.0
